ENH: Download binary data inputs with REST.
Download the binary data inputs from the Midas server and cache
locally.  A manifest of filename, md5sum is stored in a json file.
They are downloaded into the "Data" directory, and their hash checked
which the downloaddata.py script is executed or imported as a module.

The foldermd5sums.py script can be used to conveniently regenerate the
manifest.

'import downloaddata' script is added to the notebooks that need data.