Adding a testing framework based on pytest to the repository.
Adding a testing framework to improve maintainability of the
notebook repository. The testing code relies on the pytest package. We
deal with three types of errors:
1. Unexpected errors - something is broken in our code.
2. Allowed errors - errors that may or may not occur, such as calling
sitk.Show() when Fiji/ImageJ is not installed.
3. Expected errors - errors that should occur. These are errors we
intentionally introduced to illustrate a concept.

The latter two are marked in the metadata of the cell that generates
the error. For specifics see the test_notebooks.py script.