Enable testing in a memory constrained environment (i.e. CircleCI).
To enable testing in a memory constrained environment without changing
the code we decorate the sitk.ReadImage function with a shrink
filter. If the environment variable
SIMPLE_ITK_MEMORY_CONSTRAINED_ENVIRONMENT exists we assign the
decorated function back to sitk.ReadImage. This way anytime the
sitk.ReadImage function is called it either invokes the original
function (memory is not an issue) or the decorated function (memory is
an issue).