Updating notebook on cell segmentation.
Cleaned the notebook so it doesn't contain any data. Added the md5
hash of the image used in the new notebook to the manifest file. Added
the test hook for the notebook and updated the dictionary to include
the British/Australian spelling of words so that it passes the
spell-check.