Modified continuous integration script, moving from CircleCI1.0 to CircleCI2.0.
As part of the move from CircleCI1.0 to 2.0 we also no longer use
tox. Instead we have individual tests for each Python version so that
they can be run in parallel on multiple containers vs. sequentially
using tox on a single container. The tox.ini file is kept for use in
local testing.