Removed the resampling decorator, using full resolution images.
I use hard coded slice indexes when selecting a slice from the
volumes after data loading. When the
SIMPLE_ITK_MEMORY_CONSTRAINED_ENVIRONMENT environment variable is set
the images are resampled after loading. At this point the hard coded
indexs are not valid. This commit resolves the issue by using the full
resolution images, so that the indexes are valid.