BUG: segfault during display when using GetArrayViewFromImage.
The use of the numpy array from GetArrayViewFromImage is valid as long as the
SimpleITK image object hasn't been deleted. In the case of the myshow
function this is not necessarily true, so we revert to the
GetArrayFromImage function.

The bug in myshow/myshow3d was reported by Elvis Chen
(Robarts Research Institute).