BUG: intermittent error displaying image.
The code illustrating user selected/forced pixel types when reading an image causes
an intermittent error. When the type selected is sitk.sitkVectorUInt16 matplotlib
throws an exception as it doesn't support this type for display. Using the interact
method to create the dropdown intermittently sets the selection to this type. Thus
once in a while the exception is thrown, so we now allow it using the
simpleitk_error_allowed mechanism.