Updating data augmentation notebook.
Adding missing documentation for the Otsu based cropping.
Adding more details with respect to the options for the creating the
reference domain. It can either be arbitrary sized (useful when you
are constrained by a pretrained model) but then pixel spacing is
likely non-isotropic. On the other hand you can force isotropic pixel
sizes and select one of the axis sizes which determines all other axis sizes.