Changing order of download attempts for explicit url.
Previously when the manifest included an explicit url, that was the
first address we attempted to donwload from. This resulted in slow
downloads and failed CI tests due to timeouts. We now start the
download attemptes from our data stores that have faster connections
and if these fail we try the explicit url.