Updating the multi-scale demons framework.
Updated the smooth_and_resample function so that it can receive
shrink factors and smoothing sigmas per axis. This flexibility is
required when the image sizes or image spacings vary considerably
with respect to each other (e.g. in microscopy where an image
can have a size of 10K x 10K x 15).