Update ROIDataAcquisition GUI to work with 2D images too.
Previously the ROI acquisition gui only worked with 3D images, now the
UI changes based on the image dimensionality. For 3D there are two
extra sliders that control the display of the current slice and the
extent of the ROI in the axial direction. These are not visible when
dealing with a 2D image.