Updating the trust-but-verify notebook.
Needed to update the notebook and accompanying script so that the
multiprocessing also works in the notebook, and on windows replaced it
with iteration as the support on that platform is a bit problematic.
Additionally, creating symbolic links on windows required admin
permissions, so now files are copied and not linked.