Address memory issues with CI.
The 03_Image_Details notebook takes too much memory in CI causing the
kernel to intermittently die during execution. This is addressed by
running the setup_for_testing script in the first cell.

This also exposed an issue with the setup_for_testing script, it
previously did not copy the metadata dictionary of the image returned
by the original function.