Skip testing of 03_Image_Details when run via GITHUB_ACTIONS.
Executing notebook using nbconvert fails intermittently with dead
kernel when executed on a GitHub actions runner. The likely problem
was out of memory issue. Changing the notebook to use less memory did
not solve the problem. Also, when trying to debug (connecting via ssh
using the tmate action, mxschmitt/action-tmate) connected to the
runner and had the following experience: (1) Ran nbconvert on the
notebook and the kernel died. (2) Checked the memory setup, free -h,
looked like there is enough memory 3.2GiB (3) Ran nbconvert again and
this time it succeeded.