Improve the make_isotropic code, make more generic.
When creating the product of the image's two antipodal indexes, the
minimal index was explicitly listed as [0,0,0]. In combination with
the image size and the usage of the zip operator this yielded the
expected output [0,x_size],[0,y_size] and if a 3D image
[0,z_size]. This relied on the zip operator's implementation which
returns the shortest combination of inputs. Modified the code so that
it explicitly uses the image dimension to define the minimal
index. This makes the code more generic, support nD images.