Adding notebook 11_Progress into the testing, but always skipped.
Notebook 11 times out when run with nbconvert due to javascript
issues. Previously it was just commented out. Now added into the
testing framework but always skipped. The reason for this is that we
externally test all notebooks, globing on *.ipynb and if a notebook is
not specified in the tests the command fails. Now we explicitly skip
it in the tests even if it is invoked.