Change the way static C++ class methods are called in Python.
With the move to SWIG 4.1.0 calls like
sitk.ImageSeriesReader_GetGDCMSeriesIDs to static methods are not
recommended and require the usage of the -flatstaticmethod option. Code
was changed to the recommended approach, sitk.ImageSeriesReader.GetGDCMSeriesIDs

The SWIG docs (https://www.swig.org/Doc4.1/Python.html#Python_nn20):
"The last technique is only available when using the -flatstaticmethod option. This option is not recommended, it is only available for backwards compatibility as ancient versions of Python did not have Python class methods."