BUG: On windows, default delete parameter value of NamedTemporaryFile.
On windows, the delete parameter for the NamedTemporaryfile
constructor needs to be set to False, otherwise the file
is kept open and cannot be read from, default is True. If we need to
read from it, which we do, it should be set to False. It works with
delete=True on all other operating systems. The solution is to set the
value to os.name != "nt".