Change the way image sizes and intensity ranges are visualized.
Replace the usage of a histogram to visualize min and max intensity
values with a scatterplot. In the previous approach the two quantities
were disjoint. With a scatterplot it is clear what the min and max
value for the same image is.

Also, replace the use of a bar plot for image sizes to using a
scatterplot. It is easier to identify outliers in the scatterplot.