Improve parallelization and user feedback.
Switch from the multiprocess package to use concurrent.futures.

Introduce the usage of the tqdm package for displaying progress when
used in interactive mode. There is an option to disable the progress
reporting if using a scheduling system on a cluster and there is no
need for visual dispaly.

Add scatterplot output of image pixel sizes.