Set ITK to be single threaded in the characterize_data script.
The characterize_data script uses multi-processing for
concurrency. ITK filters use multi-threading as part of each process. There
is the potential for contention when too many threads are created as
compared to the number of available cores on the CPU. The solution
here is to set ITK to run in single threaded mode and the speedup is
obtained from the parallel processes.