Add support for Google Colab
Adding the option to launch notebooks using Google Colab. Unlike
Binder, colab launches each of the notebooks in an independent
environment.

To set up the notebook dependencies, the first cell clones
the git repository and installs the dependencies from the
requirements.txt file. The working directory is changed to the
Python directory so that the notebook has access to modules,
.py files, used by more than one notebook.